pulled selector into a separate component
