+++ src/Components/OpenModelControl.jsx
@@ -8,6 +8,7 @@ import TextField from '@mui/material/TextField'
 import useTheme from '@mui/styles/useTheme'
 import Dialog from './Dialog'
 import {TooltipIconButton} from './Buttons'
+import Selector from './Selector'
 import OpenIcon from '../assets/icons/Open.svg'
 import UploadIcon from '../assets/icons/Upload.svg'
 import {handleBeforeUnload} from '../utils/event'
@@ -64,6 +65,11 @@ function OpenModelDialog({isDialogDisplayed, setIsDialogDisplayed, fileOpen}) {
     fileOpen()
     setIsDialogDisplayed(false)
   }
+  const list = [
+    {0: 'one'},
+    {1: 'two'},
+    {2: 'three'},
+  ]
   return (
     <Dialog
       icon={<OpenIcon/>}
@@ -85,17 +91,17 @@ function OpenModelDialog({isDialogDisplayed, setIsDialogDisplayed, fileOpen}) {
           <p>Visit our {' '}
             <a
               target="_blank"
-              href='https://github.com/bldrs-ai/Share/wiki/Open-IFC-model-hosted-on-GitHub'
+              href='https://github.com/bldrs-ai/Share/wiki/GitHub-model-hosting'
               rel="noreferrer"
             >
               wiki
             </a> to learn more about GitHub hosting.
           </p>
           <Box>
-            <GithubOrganizationSelector setIsDialogDisplayed={setIsDialogDisplayed} label={'Organization'}/>
-            <GithubOrganizationSelector setIsDialogDisplayed={setIsDialogDisplayed} label={'Repository'}/>
-            <GithubOrganizationSelector setIsDialogDisplayed={setIsDialogDisplayed} label={'File'}/>
-            <GithubOrganizationSelector setIsDialogDisplayed={setIsDialogDisplayed} label={'Branch'}/>
+            <Selector label={'Organization'} list={list}/>
+            <Selector label={'Repository'} list={list}/>
+            <Selector label={'File'} list={list}/>
+            <Selector label={'Branch'} list={list}/>
           </Box>
           <p>Models opened from local drive cannot yet be saved or shared.</p>
         </Box>
@@ -165,81 +171,6 @@ function SampleModelFileSelector({setIsDialogDisplayed}) {
       onChange={(e) => handleSelect(e, () => setIsDialogDisplayed(false))}
       variant='outlined'
       label='Sample Projects'
-      select
-      size='small'
-    >
-      <MenuItem value={1}><Typography variant='p'>Momentum</Typography></MenuItem>
-      <MenuItem value={2}><Typography variant='p'>Schneestock</Typography></MenuItem>
-      <MenuItem value={3}><Typography variant='p'>Eisvogel</Typography></MenuItem>
-      <MenuItem value={4}><Typography variant='p'>Seestrasse</Typography></MenuItem>
-      <MenuItem value={0}><Typography variant='p'>Schependomlaan</Typography></MenuItem>
-      <MenuItem value={5}><Typography variant='p'>Structural Detail</Typography></MenuItem>
-    </TextField>
-  )
-}
-
-
-/**
- * @property {Function} setIsDialogDisplayed callback
- * @return {React.ReactElement}
- */
-function GithubOrganizationSelector({setIsDialogDisplayed, label}) {
-  const navigate = useNavigate()
-  const [selected, setSelected] = useState('')
-  const theme = useTheme()
-  const handleSelect = (e, closeDialog) => {
-    setSelected(e.target.value)
-    const modelPath = {
-      0: '/share/v/gh/IFCjs/test-ifc-files/main/Schependomlaan/IFC%20Schependomlaan.ifc#c:60.45,-4.32,60.59,1.17,5.93,-3.77',
-      1: '/share/v/gh/Swiss-Property-AG/Momentum-Public/main/Momentum.ifc#c:-38.64,12.52,35.4,-5.29,0.94,0.86',
-      2: '/share/v/gh/Swiss-Property-AG/Schneestock-Public/main/ZGRAGGEN.ifc#c:80.66,11.66,-94.06,6.32,2.93,-8.72',
-      3: '/share/v/gh/Swiss-Property-AG/Eisvogel-Public/main/EISVOGEL.ifc#c:107.36,8.46,156.67,3.52,2.03,16.71',
-      4: '/share/v/gh/Swiss-Property-AG/Seestrasse-Public/main/SEESTRASSE.ifc#c:119.61,50.37,73.68,16.18,11.25,5.74',
-      // eslint-disable-next-line max-len
-      5: '/share/v/gh/sujal23ks/BCF/main/packages/fileimport-service/ifc/ifcs/171210AISC_Sculpture_brep.ifc/120010/120020/120023/4998/2867#c:-163.46,16.12,223.99,12.03,-28.04,-15.28',
-    }
-    window.removeEventListener('beforeunload', handleBeforeUnload)
-    navigate({pathname: modelPath[e.target.value]})
-    closeDialog()
-  }
-
-  return (
-    <TextField
-      sx={{
-        'width': '260px',
-        'padding': '0px 0px 16px 0px',
-        '& .MuiOutlinedInput-input': {
-          color: theme.palette.secondary.main,
-        },
-        '& .MuiInputLabel-root': {
-          color: theme.palette.secondary.main,
-        },
-        '& .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline': {
-          borderColor: theme.palette.secondary.main,
-        },
-        '&:hover .MuiOutlinedInput-input': {
-          color: theme.palette.secondary.main,
-        },
-        '&:hover .MuiInputLabel-root': {
-          color: theme.palette.secondary.main,
-        },
-        '&:hover .MuiOutlinedInput-root .MuiOutlinedInput-notchedOutline': {
-          borderColor: theme.palette.secondary.main,
-        },
-        '& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-input': {
-          color: theme.palette.secondary.main,
-        },
-        '& .MuiInputLabel-root.Mui-focused': {
-          color: theme.palette.secondary.main,
-        },
-        '& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline': {
-          borderColor: theme.palette.secondary.main,
-        },
-      }}
-      value={selected}
-      onChange={(e) => handleSelect(e, () => setIsDialogDisplayed(false))}
-      variant='outlined'
-      label={label}
       select
       size='small'
     >
